feat(services, websocket): introduce `ServicesContext` with WebSocket support for real-time state updates
- Add `ServicesContext` to manage service-related states and operations across the app.
- Replace `CentralWebSocket` with context-based `useServicesWebSocket` for streamlined WebSocket integration.
- Refactor services page, sidebar, and terminal to use the new context for consistent state management.
- Remove direct service API calls in favor of centralized WebSocket-driven updates.
