+++ dashboard/src/components/sidebar-data-provider.tsx
@@ -4,6 +4,7 @@ import { createContext, useContext, useEffect, useState, ReactNode, useCallback 
 import { usePathname } from 'next/navigation';
 import { Group } from '@/types/groups';
 import { Service } from '@/types/services';
+import { useServices } from '@/contexts/ServicesContext';
 
 interface SidebarDataContextType {
   groups: Group[];
@@ -20,13 +21,14 @@ export function SidebarDataProvider({ children }: { children: ReactNode }) {
   const isLoginPage = pathname === '/login';
   const shouldHideSidebar = isOnboardingPage || isLoginPage;
 
+  // Use services from the global ServicesContext
+  const { services, isLoading: servicesLoading } = useServices();
+
   const [sidebarData, setSidebarData] = useState<{
     groups: Group[];
-    services: Service[];
     isLoading: boolean;
   }>({
     groups: [],
-    services: [],
     isLoading: true
   });
 
@@ -34,17 +36,11 @@ export function SidebarDataProvider({ children }: { children: ReactNode }) {
 
   const loadSidebarData = useCallback(async () => {
     try {
-      const [groupsResponse, servicesResponse] = await Promise.all([
-        fetch('/api/groups/list'),
-        fetch('/api/services/list')
-      ]);
-
+      const groupsResponse = await fetch('/api/groups/list');
       const groups = groupsResponse.ok ? await groupsResponse.json() : [];
-      const services = servicesResponse.ok ? await servicesResponse.json() : [];
 
       const newData = {
         groups: Array.isArray(groups) ? groups : [],
-        services: Array.isArray(services) ? services : [],
         isLoading: false
       };
 
@@ -52,7 +48,6 @@ export function SidebarDataProvider({ children }: { children: ReactNode }) {
 
       if (isClient) {
         localStorage.setItem('sidebarGroups', JSON.stringify(newData.groups));
-        localStorage.setItem('sidebarServices', JSON.stringify(newData.services));
       }
     } catch {
       setSidebarData(prev => ({ ...prev, isLoading: false }));
@@ -64,12 +59,10 @@ export function SidebarDataProvider({ children }: { children: ReactNode }) {
 
     try {
       const savedGroups = localStorage.getItem('sidebarGroups');
-      const savedServices = localStorage.getItem('sidebarServices');
       
-      if (savedGroups && savedServices) {
+      if (savedGroups) {
         setSidebarData({
           groups: JSON.parse(savedGroups),
-          services: JSON.parse(savedServices),
           isLoading: false
         });
       } else {
@@ -97,18 +90,11 @@ export function SidebarDataProvider({ children }: { children: ReactNode }) {
 
   const refreshData = async () => {
     try {
-
-      const [groupsResponse, servicesResponse] = await Promise.all([
-        fetch('/api/groups/list'),
-        fetch('/api/services/list')
-      ]);
-
+      const groupsResponse = await fetch('/api/groups/list');
       const groups = groupsResponse.ok ? await groupsResponse.json() : [];
-      const services = servicesResponse.ok ? await servicesResponse.json() : [];
 
       const newData = {
         groups: Array.isArray(groups) ? groups : [],
-        services: Array.isArray(services) ? services : [],
         isLoading: false
       };
 
@@ -116,11 +102,11 @@ export function SidebarDataProvider({ children }: { children: ReactNode }) {
 
       if (isClient) {
         localStorage.setItem('sidebarGroups', JSON.stringify(newData.groups));
-        localStorage.setItem('sidebarServices', JSON.stringify(newData.services));
       }
     } catch (error) {
-        console.warn('Sidebar error in sidebar-data-provider:', error);
-      }};
+      console.warn('Sidebar error in sidebar-data-provider:', error);
+    }
+  };
 
   if (shouldHideSidebar) {
     return <>{children}</>;
@@ -129,8 +115,8 @@ export function SidebarDataProvider({ children }: { children: ReactNode }) {
   return (
     <SidebarDataContext.Provider value={{
       groups: sidebarData.groups,
-      services: sidebarData.services,
-      isLoading: sidebarData.isLoading,
+      services: services, // Use services from ServicesContext
+      isLoading: sidebarData.isLoading || servicesLoading,
       refreshData
     }}>
       {children}
